requirement.txt
webbrowser
opencv-python
spotipy
fer
Pillow  
ttkbootstrap   
tkinter  ttk
python-dotenv 
